+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryAndroidTest.kt
@@ -21,6 +21,7 @@ import io.sentry.SentryOptions
 import io.sentry.SentryOptions.BeforeSendCallback
 import io.sentry.Session
 import io.sentry.ShutdownHookIntegration
+import io.sentry.SpotlightIntegration
 import io.sentry.SystemOutLogger
 import io.sentry.UncaughtExceptionHandlerIntegration
 import io.sentry.android.core.cache.AndroidEnvelopeCache
@@ -412,7 +413,7 @@ class SentryAndroidTest {
         fixture.initSut(context = mock<Application>()) { options ->
             optionsRef = options
             options.dsn = "https://key@sentry.io/123"
-            assertEquals(19, options.integrations.size)
+            assertEquals(20, options.integrations.size)
             options.integrations.removeAll {
                 it is UncaughtExceptionHandlerIntegration ||
                     it is ShutdownHookIntegration ||
@@ -431,7 +432,8 @@ class SentryAndroidTest {
                     it is SystemEventsBreadcrumbsIntegration ||
                     it is NetworkBreadcrumbsIntegration ||
                     it is TempSensorBreadcrumbsIntegration ||
-                    it is PhoneStateBreadcrumbsIntegration
+                    it is PhoneStateBreadcrumbsIntegration ||
+                    it is SpotlightIntegration
             }
         }
         assertEquals(0, optionsRef.integrations.size)